+++ test/models/concerns/folio/jw_player/file_processing_test.rb
@@ -11,7 +11,7 @@ class Folio::JwPlayer::FileProcessingTest < ActiveSupport::TestCase
   end
 
   setup do
-    @tv_file = TestMediaFile.new
+    @tv_file = TestMediaFile.new(site: get_any_site)
     @tv_file.file = Folio::Engine.root.join("test/fixtures/folio/test_7secs.mp4")
 
     assert tv_file.unprocessed?