+++ test/models/folio/belongs_to_site_test.rb
@@ -17,44 +17,38 @@ class Folio::BelongsToSiteTest < ActiveSupport::TestCase
   end
 
   test "validate_belongs_to_site" do
-    folio_site_is_a_singleton = Rails.application.config.folio_site_is_a_singleton
-    folio_site_validate_belongs_to_namespace = Rails.application.config.folio_site_validate_belongs_to_namespace
-
-    Rails.application.config.folio_site_is_a_singleton = false
-    Rails.application.config.folio_site_validate_belongs_to_namespace = true
-
     assert site = Custom::Site.create!(title: "Custom::Site",
                                        email: "custom@site.site",
                                        locale: "cs",
                                        locales: ["cs"])
 
     custom_page = Custom::Page.new(title: "Custom::Page")
-    assert_not custom_page.valid?
 
-    assert_equal :site, custom_page.errors.first.attribute
-    assert_equal :blank, custom_page.errors.first.type
-
-    custom_page.site = site
-
-    assert custom_page.valid?
-
+    # not `Custom`` namespace (as site is `Custom::Site`)
     other_page = Other::Page.new(title: "Other::Page")
-    assert_not other_page.valid?
 
-    assert_equal :site, other_page.errors.first.attribute
-    assert_equal :blank, other_page.errors.first.type
+    Rails.application.config.stub(:folio_site_validate_belongs_to_namespace, true) do
+      assert_not custom_page.valid?
+      assert_equal :site, custom_page.errors.first.attribute
+      assert_equal :blank, custom_page.errors.first.type
 
-    other_page.site = site
+      custom_page.site = site
 
-    assert_not other_page.valid?
-    assert_equal :base, other_page.errors.first.attribute
-    assert_equal :wrong_namespace, other_page.errors.first.type
+      assert custom_page.valid?
 
-    Rails.application.config.folio_site_validate_belongs_to_namespace = false
+      assert_not other_page.valid?
+      assert_equal :site, other_page.errors.first.attribute
+      assert_equal :blank, other_page.errors.first.type
 
-    assert other_page.valid?
+      other_page.site = site
 
-    Rails.application.config.folio_site_is_a_singleton = folio_site_is_a_singleton
-    Rails.application.config.folio_site_validate_belongs_to_namespace = folio_site_validate_belongs_to_namespace
+      assert_not other_page.valid?
+      assert_equal :base, other_page.errors.first.attribute
+      assert_equal :wrong_namespace, other_page.errors.first.type
+    end
+
+    Rails.application.config.stub(:folio_site_validate_belongs_to_namespace, false) do
+      assert other_page.valid?
+    end
   end
 end