+++ test/models/folio/page_singleton_test.rb
@@ -3,6 +3,8 @@
 require "test_helper"
 
 class Folio::PageSingletonTest < ActiveSupport::TestCase
+  attr_reader :site
+
   class FirstSingleton < Folio::Page
     include Folio::Singleton
   end
@@ -20,24 +22,24 @@ class Folio::PageSingletonTest < ActiveSupport::TestCase
   test "can only have one" do
     create_and_host_site
 
-    assert FirstSingleton.create!(title: "foo")
+    assert FirstSingleton.create!(title: "foo", site:)
 
     assert_equal("foo", FirstSingleton.instance.title)
     assert FirstSingleton.instance.update!(title: "oof"), "can update"
     assert_equal("oof", FirstSingleton.instance.title, "can update")
 
     assert_raises(ActiveRecord::RecordInvalid) do
-      FirstSingleton.create!(title: "bar")
+      FirstSingleton.create!(title: "bar", site:)
     end
 
-    assert SecondSingleton.create!(title: "baz")
+    assert SecondSingleton.create!(title: "baz", site:)
     assert_raises(ActiveRecord::RecordInvalid) do
-      SecondSingleton.create!(title: "bax")
+      SecondSingleton.create!(title: "bax", site:)
     end
   end
 
   test "cannot be destroyed" do
-    assert FirstSingleton.create!(title: "foo")
+    assert FirstSingleton.create!(title: "foo", site: get_any_site)
     assert_raises(ActiveRecord::RecordNotDestroyed) do
       FirstSingleton.instance.destroy!
     end