+++ CHANGELOG.md
@@ -1,5 +1,12 @@
+
 # Change Log
 All notable changes to this project will be documented in this file.
+
+## 2023-11-24
+### Changed
+- ignore `Rails.application.config.folio_site_is_a_singleton` and handle one site application as multisite with one (and main) site
+### Removed
+- Folio::Subscribable concern
 
 ## 2023-11-06
 ### Changed